fixed bug that wasnt looping each gold record since try-except was wrapping the loop
